Make the JobManager thread a daemon
The JobManager threas was not set up as a daemon, so it was not possible
to shut down JobManager or non-redis tango-rest with SIGINT or SIGTERM.

standalone JobManager has no main thread, so it can't just run a daemon
thread.

Update JobManager so the constructor does not launch the thread.
A new start method launches the thread, and the run method
runs the loop in the foreground.