Check filename and md5 hash on file uploads
Solves "caching" issue by only rejecting file uploads if the same file exists on the server **with the same filename**. Note that this also changes the REST API spec for /open from returning a list of dictionaries to returning just a simple dictionary. This will require a change on frontend that @ymzong is implementing.